move processors to processor package
