+++ worker/src/main/java/gov/cms/ab2d/worker/processor/JobCancelledException.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.worker.service;
+package gov.cms.ab2d.worker.processor;
 
 public class JobCancelledException extends RuntimeException {
